[ipq806x_retail] Use RAMFS by default
Currently we test IPQ806x on a Rumi setup, which doesn't have a flash
This change will cause the default config to generate a RAMFS
so the user do not need to modify the configuration before building
a QSDK firmware.
We will revert it back after SoD when the flash controller
will be up and running.

Change-Id: If18b84bebd41f5bfc6731e66a1b9884ec8685d9c
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>