[ipq806x_enterprise] Add qca-nss-connmgr as a loadable package
We want to make this package availabe to user on an opt-in basis
in the enterprise profile. However, we don't want to enable it by
default. Therefore, we'll use the same mechanism as in the carrier
profile, where a package is present and can be postloaded to the
board on a per-needed basis.

Change-Id: Ic009e20150c2a1f6ee7de9f13a804cd4337e1233
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>