[ipq806x-*] Add debug symbols to all builds
This change adds all the symbols required to facilitate debugging on
the product. The build will actually happen with all symbols enabled
but they'll be stripped before being installed on the FS.

Therefore, the image size will not change, but the symbols will be
accessible for offline debugging when necessary.

The kernel symbols file will be saved in kernel-debug.tar.bz2 as per
the standard openwrt mechanism. User symbols are not saved yet as they
will require additional changes to the framework.

Change-Id: I0890cf08eb98a617bca0f8360fc8ba513784343a
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>